refactor: use weaviate.ts instead of weaviateHelpers.ts
